+++ branch-master/sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2026-08-01T01:16:07+0000
+// 2026-09-01T03:17:11+0000
 (function () {
 "use strict";