+++ branch-master/CHANGELOG.md
@@ -7,6 +7,15 @@
 - sqlite - Add similar error-handling from builtin-sql-function PERCENTILE() into custom-sql-function QUANTILE().
 - none
 
+# v2026.8.31
+- jslint-ci - Update ci-shell-function shGitLsTree() to include sha256 hash of files.
+- sqlmath - Cleanup c-function winSinefitSnr().
+- sqlmath - Update sql-function WIN_SINEFIT2() argument modeSnr with bitmask, allowing selective sine-fit of each column, to avoid wasting cpu on columns that don't need it.
+- lgbm - Upgrade to lightgbm-v4.7.0.
+- sqlite - Upgrade to sqlite-v3.51.3.
+- sqlmath - Remove obsolete sql-functions MEDIAN2(), QUANTILE(), superseded by sqlite-builtins.
+- sqlmath - Fix c-header conflict between dlfcn.h and Python.h.
+
 # v2026.7.30
 - jslint-ecma - Add ES2015-feature for..of.
 - jslint - Change scope from scope_function to scope_block: